+++ src/problems/problem_037.rs
@@ -16,7 +16,7 @@ pub fn solve() -> u32 {
     // Primes can only end in 1, 3, 7, or 9
     for y in [1, 3, 7, 9] {
       let candidate = current * 10 + y; // Add digit at the end
-      if is_prime(candidate as u64) {
+      if is_prime(candidate) {
         queue.push_back(candidate);
         right_truncatable.insert(candidate);
       }
@@ -32,7 +32,7 @@ pub fn solve() -> u32 {
     // Here we must test all digits
     for y in [1, 2, 3, 4, 5, 6, 7, 8, 9] {
       let candidate = y * 10_u32.pow(current.ilog10() + 1) + current; // Add digit in front
-      if candidate <= *maximum && is_prime(candidate as u64) {
+      if candidate <= *maximum && is_prime(candidate) {
         queue.push_back(candidate);
         left_truncatable.insert(candidate);
       }
